+++ fnf-20190315/alibabacloud_fnf20190315/client.py
@@ -288,11 +288,11 @@ class Client(OpenApiClient):
         @return: DeleteFlowResponse
         """
         UtilClient.validate_model(request)
-        query = {}
+        body = {}
         if not UtilClient.is_unset(request.name):
-            query['Name'] = request.name
+            body['Name'] = request.name
         req = open_api_models.OpenApiRequest(
-            query=OpenApiUtilClient.query(query)
+            body=OpenApiUtilClient.parse_to_map(body)
         )
         params = open_api_models.Params(
             action='DeleteFlow',
@@ -324,11 +324,11 @@ class Client(OpenApiClient):
         @return: DeleteFlowResponse
         """
         UtilClient.validate_model(request)
-        query = {}
+        body = {}
         if not UtilClient.is_unset(request.name):
-            query['Name'] = request.name
+            body['Name'] = request.name
         req = open_api_models.OpenApiRequest(
-            query=OpenApiUtilClient.query(query)
+            body=OpenApiUtilClient.parse_to_map(body)
         )
         params = open_api_models.Params(
             action='DeleteFlow',
@@ -380,13 +380,13 @@ class Client(OpenApiClient):
         runtime: util_models.RuntimeOptions,
     ) -> fnf_20190315_models.DeleteScheduleResponse:
         UtilClient.validate_model(request)
-        query = {}
+        body = {}
         if not UtilClient.is_unset(request.flow_name):
-            query['FlowName'] = request.flow_name
+            body['FlowName'] = request.flow_name
         if not UtilClient.is_unset(request.schedule_name):
-            query['ScheduleName'] = request.schedule_name
+            body['ScheduleName'] = request.schedule_name
         req = open_api_models.OpenApiRequest(
-            query=OpenApiUtilClient.query(query)
+            body=OpenApiUtilClient.parse_to_map(body)
         )
         params = open_api_models.Params(
             action='DeleteSchedule',
@@ -410,13 +410,13 @@ class Client(OpenApiClient):
         runtime: util_models.RuntimeOptions,
     ) -> fnf_20190315_models.DeleteScheduleResponse:
         UtilClient.validate_model(request)
-        query = {}
+        body = {}
         if not UtilClient.is_unset(request.flow_name):
-            query['FlowName'] = request.flow_name
+            body['FlowName'] = request.flow_name
         if not UtilClient.is_unset(request.schedule_name):
-            query['ScheduleName'] = request.schedule_name
+            body['ScheduleName'] = request.schedule_name
         req = open_api_models.OpenApiRequest(
-            query=OpenApiUtilClient.query(query)
+            body=OpenApiUtilClient.parse_to_map(body)
         )
         params = open_api_models.Params(
             action='DeleteSchedule',